refactor: split landing.rs into focused modules (#202)
Mechanically split the 2,068-line frontend/src/pages/landing.rs into a thin
module root (LandingPage + short_path) plus six submodules under
frontend/src/pages/landing/: sorting, filtering (pure sort/filter/facet helpers
+ their tests), grid, toolbar, filters, and table (the inline-edit components).
Feature-gating for the is_admin signal is preserved verbatim in LandingPage so
SSR and WASM still render identical markup. No behavior change.
