+++ frontend/src/pages/landing.rs
@@ -1,37 +1,22 @@
-use std::cmp::Ordering;
-use std::collections::{BTreeMap, HashSet};
-
 use dioxus::prelude::*;
-use dioxus_router::use_navigator;
-use omnibus_shared::{
-    Contributor, EbookLibrary, EbookMetadata, MetadataOverrides, SortDir, SortKey, TagWeight,
-    ViewFilters, ViewMode, ViewPrefs,
-};
+use omnibus_shared::{EbookLibrary, SortKey, TagWeight, ViewFilters, ViewMode, ViewPrefs};
 
-use crate::components::chip_editor::{collect_suggestions, ChipEditor, SuggestionItem};
-use crate::{data, use_search_query, use_server_url, view_prefs, Route};
+use crate::components::chip_editor::{collect_suggestions, SuggestionItem};
+use crate::{data, use_search_query, use_server_url, view_prefs};
 
-/// Inline-editable field on a power-user-table row (F5.9-lite admin
-/// surface). Used by `EbookRow` to track which of its cells (if any) is
-/// in edit mode at any given time — single-cell-at-a-time keeps the
-/// keyboard/blur lifecycle simple and avoids fighting the row's
-/// click-to-navigate handler.
-///
-/// `Series` edits the series *name* only — the series index ("#N")
-/// stays in the F5.1 metadata edit page on `/books/:uuid/edit` for v1
-/// to avoid adding a separate Series Index column to the power-user
-/// table. The user's stated cleanup pain (stripping series-prefix
-/// cruft from titles, renaming author/series variants) is satisfied
-/// by Title + Series + Authors.
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-enum EditField {
-    Title,
-    Series,
-    Publisher,
-    Published,
-    Language,
-    Authors,
-}
+mod filtering;
+mod filters;
+mod grid;
+mod sorting;
+mod table;
+mod toolbar;
+
+use filtering::{apply_filters, facet_counts};
+use filters::{EmptyFiltered, FilterSidebar, FormatChips};
+use grid::BookGrid;
+use sorting::{default_dir_for, sort_books, toggle_dir};
+use table::BookTable;
+use toolbar::Toolbar;
 
 /// Landing page — primary library surface.
 ///
@@ -305,1764 +290,14 @@ fn short_path(path: &str) -> String {
         .to_string()
 }
 
-// ---------------------------------------------------------------------------
-// Toolbar
-// ---------------------------------------------------------------------------
-
-#[component]
-fn Toolbar(prefs: ViewPrefs, on_change: EventHandler<ViewPrefs>) -> Element {
-    let view_mode = prefs.view_mode;
-    let sort_key = prefs.sort_key;
-    let sort_dir = prefs.sort_dir;
-    let filters_open = prefs.filters_open;
-
-    let apply = move |new_prefs: ViewPrefs| on_change.call(new_prefs);
-    let set_view = {
-        let prefs = prefs.clone();
-        move |mode: ViewMode| {
-            let mut next = prefs.clone();
-            next.view_mode = mode;
-            apply(next);
-        }
-    };
-    let toggle_filters = {
-        let prefs = prefs.clone();
-        move |_| {
-            let mut next = prefs.clone();
-            next.filters_open = !next.filters_open;
-            apply(next);
-        }
-    };
-    let set_sort_key = {
-        let prefs = prefs.clone();
-        move |key: SortKey| {
-            let mut next = prefs.clone();
-            // Switching to a different axis from the grid dropdown should
-            // adopt that axis's natural direction (descending for time-based
-            // axes, ascending for alphabetical) — matches the table-view
-            // header behavior so the two views stay consistent.
-            if next.sort_key != key {
-                next.sort_dir = default_dir_for(key);
-            }
-            next.sort_key = key;
-            apply(next);
-        }
-    };
-    let toggle_sort_dir = {
-        let prefs = prefs.clone();
-        move |_| {
-            let mut next = prefs.clone();
-            next.sort_dir = toggle_dir(next.sort_dir);
-            apply(next);
-        }
-    };
-
-    let set_view_table = set_view.clone();
-    let set_view_grid = set_view.clone();
-
-    rsx! {
-        div { class: "lib-toolbar", role: "toolbar", "data-testid": "lib-toolbar",
-            button {
-                class: "lib-toggle-btn lib-filters-btn",
-                "aria-pressed": "{filters_open}",
-                "data-testid": "lib-filters-toggle",
-                aria_label: "Toggle filter sidebar",
-                onclick: toggle_filters,
-                "Filters"
-            }
-            // Pressed-button toggle group, not an ARIA tablist — there are no
-            // associated tab panels and no arrow-key tab navigation, so
-            // `aria-pressed` on plain `<button>`s is the right shape.
-            div { class: "lib-view-toggle", "aria-label": "View mode",
-                button {
-                    class: "lib-toggle-btn",
-                    "aria-pressed": "{view_mode == ViewMode::Table}",
-                    "data-testid": "view-toggle-table",
-                    onclick: move |_| set_view_table(ViewMode::Table),
-                    "Table"
-                }
-                button {
-                    class: "lib-toggle-btn",
-                    "aria-pressed": "{view_mode == ViewMode::Grid}",
-                    "data-testid": "view-toggle-grid",
-                    onclick: move |_| set_view_grid(ViewMode::Grid),
-                    "Grid"
-                }
-            }
-
-            if view_mode == ViewMode::Grid {
-                div { class: "lib-sort-controls",
-                    label { class: "lib-sort-label",
-                        "Sort by"
-                        select {
-                            class: "lib-sort-select",
-                            "data-testid": "lib-sort-select",
-                            onchange: move |evt: Event<FormData>| {
-                                if let Some(key) = sort_key_from_value(&evt.value()) {
-                                    set_sort_key(key);
-                                }
-                            },
-                            for opt in SORT_KEYS.iter().copied() {
-                                option {
-                                    key: "{sort_key_value(opt)}",
-                                    value: "{sort_key_value(opt)}",
-                                    selected: opt == sort_key,
-                                    "{sort_key_label(opt)}"
-                                }
-                            }
-                        }
-                    }
-                    button {
-                        class: "lib-sort-dir",
-                        "data-testid": "lib-sort-dir",
-                        aria_label: "Toggle sort direction",
-                        onclick: toggle_sort_dir,
-                        if sort_dir == SortDir::Asc { "↑" } else { "↓" }
-                    }
-                }
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Filter sidebar
-// ---------------------------------------------------------------------------
-
-#[derive(Clone, PartialEq)]
-struct FacetCounts {
-    authors: Vec<(String, usize)>,
-    series: Vec<(String, usize)>,
-    /// Normalized lowercase format keys (`"epub"`, `"m4b"`, …) paired with
-    /// counts. The display label is derived at render time via
-    /// [`format_display_label`] so the keys stay canonical.
-    formats: Vec<(String, usize)>,
-    tags: Vec<(String, usize)>,
-}
-
-#[component]
-fn FilterSidebar(
-    facets: FacetCounts,
-    filters: ViewFilters,
-    on_change: EventHandler<ViewFilters>,
-) -> Element {
-    // Reflect every filter bucket — including `formats` set via the top
-    // chip row — so a user with only a format chip selected still sees
-    // the sidebar's "Clear filters" affordance.
-    let any_active = !filters.is_empty();
-    let toggle = {
-        let filters = filters.clone();
-        move |group: &'static str, value: String| {
-            let mut next = filters.clone();
-            let bucket = match group {
-                "authors" => &mut next.authors,
-                "series" => &mut next.series,
-                "tags" => &mut next.tags,
-                other => {
-                    debug_assert!(
-                        false,
-                        "FilterSidebar::toggle: unknown facet group {other:?}"
-                    );
-                    return;
-                }
-            };
-            if let Some(pos) = bucket.iter().position(|v| v == &value) {
-                bucket.remove(pos);
-            } else {
-                bucket.push(value);
-            }
-            on_change.call(next);
-        }
-    };
-
-    rsx! {
-        aside { class: "lib-sidebar", "data-testid": "lib-sidebar", aria_label: "Filters",
-            if any_active {
-                button {
-                    class: "lib-clear-filters",
-                    "data-testid": "lib-clear-filters",
-                    onclick: move |_| on_change.call(ViewFilters::default()),
-                    "Clear filters"
-                }
-            }
-
-            FacetSection {
-                title: "Authors",
-                testid: "lib-facet-authors",
-                items: facets.authors.clone(),
-                selected: filters.authors.clone(),
-                on_toggle: {
-                    let toggle = toggle.clone();
-                    move |v: String| toggle("authors", v)
-                },
-            }
-            FacetSection {
-                title: "Series",
-                testid: "lib-facet-series",
-                items: facets.series.clone(),
-                selected: filters.series.clone(),
-                on_toggle: {
-                    let toggle = toggle.clone();
-                    move |v: String| toggle("series", v)
-                },
-            }
-            FacetSection {
-                title: "Tags",
-                testid: "lib-facet-tags",
-                items: facets.tags.clone(),
-                selected: filters.tags.clone(),
-                on_toggle: {
-                    let toggle = toggle.clone();
-                    move |v: String| toggle("tags", v)
-                },
-            }
-        }
-    }
-}
-
-#[component]
-fn FacetSection(
-    title: String,
-    testid: String,
-    items: Vec<(String, usize)>,
-    selected: Vec<String>,
-    on_toggle: EventHandler<String>,
-) -> Element {
-    if items.is_empty() {
-        return rsx! { Fragment {} };
-    }
-    let selected_set: HashSet<&String> = selected.iter().collect();
-    rsx! {
-        section { class: "lib-facet", "data-testid": "{testid}",
-            h3 { class: "lib-facet-title", "{title}" }
-            ul { class: "lib-chip-list",
-                for (name, count) in items.iter() {
-                    li {
-                        key: "{name}",
-                        button {
-                            // Layer Atrium's `.chip` look onto the existing
-                            // `.lib-chip` class — the Playwright selector
-                            // `button.lib-chip[data-value="…"]` still matches.
-                            class: "chip lib-chip",
-                            "aria-pressed": "{selected_set.contains(&name)}",
-                            "data-value": "{name}",
-                            title: "{name}",
-                            onclick: {
-                                let name = name.clone();
-                                move |_| on_toggle.call(name.clone())
-                            },
-                            span { class: "lib-chip-label", "{name}" }
-                            span { class: "count lib-chip-count", "{count}" }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Format chips (top-of-page inline filter)
-// ---------------------------------------------------------------------------
-
-#[component]
-fn FormatChips(
-    counts: Vec<(String, usize)>,
-    visible_count: usize,
-    book_count: usize,
-    selected: Vec<String>,
-    on_change: EventHandler<Vec<String>>,
-) -> Element {
-    if counts.is_empty() {
-        return rsx! { Fragment {} };
-    }
-    let selected_set: HashSet<String> = selected.iter().cloned().collect();
-    let all_active = selected.is_empty();
-
-    rsx! {
-        div { class: "lib-format-chips",
-            "data-testid": "lib-format-chips",
-            role: "group",
-            aria_label: "Format filters",
-
-            span { class: "label lib-format-chips-label", "Filter" }
-
-            button {
-                class: if all_active { "chip on" } else { "chip" },
-                "data-format": "all",
-                "aria-pressed": "{all_active}",
-                onclick: move |_| on_change.call(Vec::new()),
-                // Count distinct books, not per-format memberships — a book
-                // with EPUB+M4B contributes 1, not 2.
-                "All formats "
-                span { class: "count", "{book_count}" }
-            }
-
-            for (key, count) in counts.into_iter() {
-                {
-                    let is_selected = selected_set.contains(&key);
-                    let label = format_display_label(&key);
-                    let key_for_click = key.clone();
-                    let selected_for_click = selected.clone();
-                    rsx! {
-                        button {
-                            key: "{key}",
-                            class: if is_selected { "chip on" } else { "chip" },
-                            "data-format": "{key}",
-                            "aria-pressed": "{is_selected}",
-                            onclick: move |_| {
-                                let mut next: Vec<String> = selected_for_click.clone();
-                                if let Some(pos) = next.iter().position(|v| v == &key_for_click) {
-                                    next.remove(pos);
-                                } else {
-                                    next.push(key_for_click.clone());
-                                }
-                                on_change.call(next);
-                            },
-                            "{label} "
-                            span { class: "count", "{count}" }
-                        }
-                    }
-                }
-            }
-
-            div { class: "lib-format-chips-spacer" }
-            span { class: "mono lib-format-chips-count",
-                "{visible_count} of {book_count}"
-            }
-        }
-    }
-}
-
-#[component]
-fn EmptyFiltered(on_clear: EventHandler<()>) -> Element {
-    rsx! {
-        div { class: "library-empty",
-            p { "No books match these filters." }
-            button {
-                class: "btn",
-                "data-testid": "lib-clear-filters-empty",
-                onclick: move |_| on_clear.call(()),
-                "Clear filters"
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Table view
-// ---------------------------------------------------------------------------
-
-#[component]
-fn BookTable(
-    books: Vec<EbookMetadata>,
-    prefs: ViewPrefs,
-    on_sort: EventHandler<SortKey>,
-    server_url: String,
-    is_admin: bool,
-    // Forwarded by reference (ReadSignal) all the way down so the
-    // suggestion pool isn't cloned at any intermediate layer — even on
-    // libraries with thousands of authors.
-    author_suggestions: ReadSignal<Vec<SuggestionItem>>,
-    tag_suggestions: ReadSignal<Vec<SuggestionItem>>,
-) -> Element {
-    rsx! {
-        div {
-            id: "ebook-table",
-            "data-testid": "ebook-table",
-            class: "ebook-table-wrap",
-            table { class: "ebook-table",
-                thead {
-                    tr {
-                        th { class: "ebook-col-cover", "Cover" }
-                        SortableHeader {
-                            class: "ebook-col-title".to_string(),
-                            label: "Title".to_string(),
-                            sort_key: SortKey::Title,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        SortableHeader {
-                            class: "ebook-col-author".to_string(),
-                            label: "Author".to_string(),
-                            sort_key: SortKey::Author,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        SortableHeader {
-                            class: "ebook-col-series".to_string(),
-                            label: "Series".to_string(),
-                            sort_key: SortKey::Series,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        th { class: "ebook-col-publisher", "Publisher" }
-                        th { class: "ebook-col-published", "Published" }
-                        th { class: "ebook-col-formats", "Formats" }
-                        SortableHeader {
-                            class: "ebook-col-updated".to_string(),
-                            label: "Last Updated".to_string(),
-                            sort_key: SortKey::LastUpdated,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        SortableHeader {
-                            class: "ebook-col-added".to_string(),
-                            label: "Added".to_string(),
-                            sort_key: SortKey::NewestAdded,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        th { class: "ebook-col-language", "Language" }
-                    }
-                }
-                tbody {
-                    for book in books.into_iter() {
-                        EbookRow {
-                            key: "{book.filename}",
-                            book: book,
-                            server_url: server_url.clone(),
-                            is_admin,
-                            author_suggestions,
-                            tag_suggestions,
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SortableHeader(
-    class: String,
-    label: String,
-    sort_key: SortKey,
-    prefs: ViewPrefs,
-    on_sort: EventHandler<SortKey>,
-) -> Element {
-    let active = prefs.sort_key == sort_key;
-    let aria_sort = match (active, prefs.sort_dir) {
-        (true, SortDir::Asc) => "ascending",
-        (true, SortDir::Desc) => "descending",
-        _ => "none",
-    };
-    let arrow = if !active {
-        ""
-    } else if prefs.sort_dir == SortDir::Asc {
-        " ↑"
-    } else {
-        " ↓"
-    };
-    rsx! {
-        th { class: "{class} sort-th", aria_sort: "{aria_sort}",
-            button {
-                class: "sort-th-btn",
-                onclick: move |_| on_sort.call(sort_key),
-                "{label}{arrow}"
-            }
-        }
-    }
-}
-
-#[component]
-fn EbookRow(
-    book: EbookMetadata,
-    server_url: String,
-    is_admin: bool,
-    author_suggestions: ReadSignal<Vec<SuggestionItem>>,
-    tag_suggestions: ReadSignal<Vec<SuggestionItem>>,
-) -> Element {
-    // Stable per-book uuid drives both the detail-route URL and the
-    // thumbnail URL — see `Route::BookDetail` for why it's keyed on the
-    // uuid instead of `books.id`.
-    let uuid = book.unique_identifier.clone().unwrap_or_default();
-    let row_testid = format!("ebook-row-{}", row_slug(&book.filename));
-    let has_cover = book.cover_url.is_some();
-    let thumb_base = format!("{server_url}/api/thumbs/{uuid}");
-    let _ = tag_suggestions; // reserved for a future inline-tag editor
-
-    // Optimistic per-row state: we render from `book_state` rather than
-    // the raw `book` prop so a successful inline save updates the row
-    // immediately, without a full library refetch. The server-side
-    // merge in `rpc_save_overrides` returns the canonical merged
-    // EbookMetadata which we install verbatim.
-    //
-    // Resync from the prop when the upstream `book` changes (library
-    // refetch / sort change reusing the same EbookRow instance via the
-    // filename key) — but only when no cell is mid-edit, so a save
-    // round-trip doesn't clobber an open input. `use_reactive!` keys
-    // the effect on `book` itself.
-    let mut book_state: Signal<EbookMetadata> = use_signal(|| book.clone());
-    let editing: Signal<Option<EditField>> = use_signal(|| None);
-    use_effect(use_reactive!(|book| {
-        if editing().is_none() {
-            book_state.set(book.clone());
-        }
-    }));
-
-    let nav = use_navigator();
-    let uuid_click = uuid.clone();
-    let uuid_key = uuid.clone();
-    let uuid_for_save = uuid.clone();
-    let url_for_save = server_url.clone();
-
-    // Common save callback used by every inline cell. Builds a sparse
-    // `MetadataOverrides`, POSTs to the existing F5.1 endpoint, then
-    // installs the server-returned merged metadata into `book_state`.
-    // Empty strings clear the override for the field (None — the
-    // scanned value re-surfaces); non-empty strings persist as
-    // `Some(value)`.
-    let save_field = move |field: EditField, value: String| {
-        let uuid = uuid_for_save.clone();
-        let url = url_for_save.clone();
-        spawn(async move {
-            let mut overrides = MetadataOverrides::default();
-            let trimmed = value.trim();
-            let value = if trimmed.is_empty() {
-                None
-            } else {
-                Some(trimmed.to_string())
-            };
-            match field {
-                EditField::Title => overrides.title = value,
-                EditField::Series => overrides.series = value,
-                EditField::Publisher => overrides.publisher = value,
-                EditField::Published => overrides.published = value,
-                EditField::Language => overrides.language = value,
-                EditField::Authors => {
-                    // Authors is handled below — this branch is reserved
-                    // for the chip-editor on_change path which sets
-                    // `creators` instead.
-                    return;
-                }
-            }
-            if overrides.validate().is_err() {
-                // Validation rejects payloads that exceed the F5.1
-                // length caps. The display reverts on the next
-                // refetch; a toast / inline error message ships in a
-                // follow-up PR.
-                return;
-            }
-            if let Ok(Some(merged)) = data::save_overrides(&url, &uuid, &overrides).await {
-                book_state.set(merged);
-            }
-        });
-    };
-    // Authors chip editor lifts the canonical creator list out of the
-    // optimistic row state into a Signal so ChipEditor can drive it
-    // directly. Edits go through `save_authors` (separate callback so
-    // it can build the `creators` override; the text-cell path is
-    // scalar-only).
-    let initial_authors: Vec<String> = book_state
-        .read()
-        .creators
-        .iter()
-        .map(|c| c.name.clone())
-        .collect();
-    let mut authors_draft: Signal<Vec<String>> = use_signal(|| initial_authors);
-    // Keep `authors_draft` in sync with `book_state.creators` so a
-    // save round-trip (which may canonicalize / reorder names on the
-    // server) doesn't leave the chip editor showing stale chips, and
-    // re-opening the cell after an external update shows the current
-    // truth. The effect is a no-op when the values already match, so
-    // the user's in-progress chip edits aren't clobbered between
-    // `on_change` and the optimistic `book_state` install.
-    use_effect(move || {
-        let canonical: Vec<String> = book_state
-            .read()
-            .creators
-            .iter()
-            .map(|c| c.name.clone())
-            .collect();
-        if *authors_draft.peek() != canonical {
-            authors_draft.set(canonical);
-        }
-    });
-    let uuid_for_authors = uuid.clone();
-    let url_for_authors = server_url.clone();
-    let save_authors = move |new_names: Vec<String>| {
-        let uuid = uuid_for_authors.clone();
-        let url = url_for_authors.clone();
-        spawn(async move {
-            let creators: Vec<Contributor> = new_names
-                .iter()
-                .map(|name| Contributor {
-                    name: name.clone(),
-                    ..Default::default()
-                })
-                .collect();
-            let overrides = MetadataOverrides {
-                creators: Some(creators),
-                ..Default::default()
-            };
-            if overrides.validate().is_err() {
-                return;
-            }
-            if let Ok(Some(merged)) = data::save_overrides(&url, &uuid, &overrides).await {
-                book_state.set(merged);
-            }
-        });
-    };
-
-    // Pull values from the optimistic state so a save's returned
-    // merged metadata is what we render.
-    let display_book = book_state.read().clone();
-    let display_title = display_book
-        .title
-        .as_deref()
-        .unwrap_or(&display_book.filename)
-        .to_string();
-    let series_line = match (
-        display_book.series.as_deref(),
-        display_book.series_index.as_deref(),
-    ) {
-        (Some(s), Some(i)) => format!("{s} #{i}"),
-        (Some(s), None) => s.to_string(),
-        _ => String::new(),
-    };
-    let authors_text = contributor_names(&display_book.creators);
-    let updated = display_book.modified.as_deref().unwrap_or("").to_string();
-    let added = display_book.added_at.as_deref().unwrap_or("").to_string();
-    let publisher = display_book.publisher.clone().unwrap_or_default();
-    let published = display_book.published.clone().unwrap_or_default();
-    let language = display_book.language.clone().unwrap_or_default();
-    let series_text = display_book.series.clone().unwrap_or_default();
-
-    rsx! {
-        tr {
-                class: "ebook-row",
-                "data-testid": "{row_testid}",
-                id: "{row_testid}",
-                role: "button",
-                tabindex: "0",
-                aria_label: "Open details for {display_title}",
-                onclick: move |_| {
-                    // Row-level click navigates only when no cell-level
-                    // edit is in progress. Each editable cell stops
-                    // propagation on click already, but a click on a
-                    // non-editable area (cover, formats, dates) while
-                    // a cell is open would otherwise blur+save the
-                    // input AND fire the navigation. Bailing on the
-                    // editing-is-some path keeps the user on the page
-                    // while the blur-save lands.
-                    if editing().is_some() {
-                        return;
-                    }
-                    nav.push(Route::BookDetail { uuid: uuid_click.clone() });
-                },
-                onkeydown: move |evt: Event<KeyboardData>| {
-                    if editing().is_some() {
-                        return;
-                    }
-                    let key = evt.key();
-                    if key == Key::Enter || key == Key::Character(" ".to_string()) {
-                        evt.prevent_default();
-                        nav.push(Route::BookDetail { uuid: uuid_key.clone() });
-                    }
-                },
-                td { class: "ebook-col-cover", "data-testid": "ebook-cell-cover",
-                    if has_cover {
-                        img {
-                            class: "ebook-thumb",
-                            src: "{thumb_base}/md",
-                            srcset: "{thumb_base}/sm 160w, {thumb_base}/md 320w, {thumb_base}/lg 640w",
-                            sizes: "(max-width: 640px) 160px, (max-width: 1280px) 320px, 640px",
-                            alt: "Cover of {display_title}",
-                            loading: "lazy",
-                            width: "320",
-                            height: "480",
-                        }
-                    } else {
-                        div { class: "ebook-thumb ebook-thumb-fallback", "—" }
-                    }
-                }
-                EditableCell {
-                    col_class: "ebook-col-title".to_string(),
-                    cell_testid: "ebook-cell-title".to_string(),
-                    field: EditField::Title,
-                    display_value: display_title.clone(),
-                    is_admin,
-                    editing,
-                    placeholder: "Title".to_string(),
-                    on_save: { let save = save_field.clone(); move |v: String| save(EditField::Title, v) },
-                    error: display_book.error.clone(),
-                }
-                AuthorsCell {
-                    is_admin,
-                    editing,
-                    authors_draft,
-                    authors_text: authors_text.clone(),
-                    suggestions: author_suggestions,
-                    on_change: {
-                        let save = save_authors;
-                        move |names: Vec<String>| save(names)
-                    },
-                }
-                EditableCell {
-                    col_class: "ebook-col-series".to_string(),
-                    cell_testid: "ebook-cell-series".to_string(),
-                    field: EditField::Series,
-                    // Display: "Pioneers #1" (name + index, F1.3 standard).
-                    // Edit:    "Pioneers"    (the series-name override is
-                    //                         scalar; index lives in
-                    //                         `series_index` which the
-                    //                         landing-table doesn't
-                    //                         currently expose for edit).
-                    display_value: series_line.clone(),
-                    edit_value: Some(series_text.clone()),
-                    is_admin,
-                    editing,
-                    placeholder: "Series".to_string(),
-                    on_save: { let save = save_field.clone(); move |v: String| save(EditField::Series, v) },
-                    error: None,
-                }
-                EditableCell {
-                    col_class: "ebook-col-publisher".to_string(),
-                    cell_testid: "ebook-cell-publisher".to_string(),
-                    field: EditField::Publisher,
-                    display_value: publisher.clone(),
-                    is_admin,
-                    editing,
-                    placeholder: "Publisher".to_string(),
-                    on_save: { let save = save_field.clone(); move |v: String| save(EditField::Publisher, v) },
-                    error: None,
-                }
-                EditableCell {
-                    col_class: "ebook-col-published".to_string(),
-                    cell_testid: "ebook-cell-published".to_string(),
-                    field: EditField::Published,
-                    display_value: published.clone(),
-                    is_admin,
-                    editing,
-                    placeholder: "YYYY-MM-DD".to_string(),
-                    on_save: { let save = save_field.clone(); move |v: String| save(EditField::Published, v) },
-                    error: None,
-                }
-                td { class: "ebook-col-formats", "data-testid": "ebook-cell-formats",
-                    if display_book.formats.is_empty() {
-                        span { class: "ebook-cell-formats-empty", "—" }
-                    } else {
-                        for fmt in display_book.formats.iter() {
-                            span { class: "format-badge", "{format_badge_label(fmt)}" }
-                        }
-                    }
-                }
-                td { class: "ebook-col-updated", "data-testid": "ebook-cell-updated", "{updated}" }
-                td { class: "ebook-col-added", "data-testid": "ebook-cell-added", "{added}" }
-                EditableCell {
-                    col_class: "ebook-col-language".to_string(),
-                    cell_testid: "ebook-cell-language".to_string(),
-                    field: EditField::Language,
-                    display_value: language.clone(),
-                    is_admin,
-                    editing,
-                    placeholder: "en".to_string(),
-                    on_save: { let save = save_field.clone(); move |v: String| save(EditField::Language, v) },
-                    error: None,
-                }
-            }
-    }
-}
-
-/// Inline-editable text cell used by [`EbookRow`]. Renders a span of
-/// text by default; in admin mode, a click swaps to a text input that
-/// commits via [`EditField::on_save`] on Enter or blur and cancels on
-/// Escape. The input's `onclick` stops propagation so the row-level
-/// navigate handler doesn't fire while the user is editing.
-///
-/// `suppress_inline_input` is the Authors-cell escape hatch: the click
-/// still toggles `editing` (so the expansion row appears) but the
-/// in-cell input never renders.
-#[component]
-fn EditableCell(
-    col_class: String,
-    cell_testid: String,
-    field: EditField,
-    display_value: String,
-    /// Separate value used to seed the input when edit mode opens.
-    /// Some cells render a richer display string than the underlying
-    /// scalar override — Series shows "Pioneers #1" but only "Pioneers"
-    /// is the series-name override. Pass `Some(bare_value)` so the
-    /// input seeds (and the blur-comparison runs against) the editable
-    /// scalar, not the rendered text. Defaults to `display_value`.
-    #[props(default)]
-    edit_value: Option<String>,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    placeholder: String,
-    on_save: EventHandler<String>,
-    error: Option<String>,
-) -> Element {
-    let mut editing = editing;
-    let mut draft = use_signal(String::new);
-    let is_editing = editing() == Some(field);
-    let active_class = if is_editing {
-        " ebook-cell-editing"
-    } else {
-        ""
-    };
-    let admin_class = if is_admin { " ebook-cell-editable" } else { "" };
-    let combined_class = format!("{col_class}{admin_class}{active_class}");
-    // Single source of truth for "what the cell holds right now" —
-    // used to seed the draft on click, restore it on Escape, and skip
-    // the save-on-blur round-trip when nothing actually changed.
-    let initial = edit_value.unwrap_or_else(|| display_value.clone());
-    let initial_for_click = initial.clone();
-    let initial_for_cancel = initial.clone();
-    let initial_for_enter = initial.clone();
-    let initial_for_blur = initial.clone();
-    let testid_input = format!("{cell_testid}-input");
-
-    rsx! {
-        td {
-            class: "{combined_class}",
-            "data-testid": "{cell_testid}",
-            onclick: move |e| {
-                if !is_admin {
-                    return;
-                }
-                e.stop_propagation();
-                if !is_editing {
-                    draft.set(initial_for_click.clone());
-                    editing.set(Some(field));
-                }
-            },
-            if is_editing {
-                input {
-                    class: "ebook-cell-edit",
-                    "data-testid": "{testid_input}",
-                    autofocus: true,
-                    value: "{draft}",
-                    placeholder: "{placeholder}",
-                    onclick: move |e| { e.stop_propagation(); },
-                    oninput: move |e| { draft.set(e.value()); },
-                    onkeydown: move |e: Event<KeyboardData>| {
-                        match e.key() {
-                            Key::Enter => {
-                                e.prevent_default();
-                                // Stop the Enter event from bubbling to
-                                // the row-level keydown that navigates
-                                // to the book detail page.
-                                e.stop_propagation();
-                                let current = draft();
-                                if current.trim() != initial_for_enter.trim() {
-                                    on_save.call(current);
-                                }
-                                editing.set(None);
-                            }
-                            Key::Escape => {
-                                e.prevent_default();
-                                e.stop_propagation();
-                                draft.set(initial_for_cancel.clone());
-                                editing.set(None);
-                            }
-                            _ => {}
-                        }
-                    },
-                    onblur: move |_| {
-                        // Save on blur only when the draft actually
-                        // changed. Clicking into a cell and clicking
-                        // back out without typing must not POST an
-                        // override equal to the scanned value (which
-                        // would leak `metadata_overrides` rows that
-                        // match the underlying scan, defeating the
-                        // F5.1 merge semantics).
-                        let current = draft();
-                        if current.trim() != initial_for_blur.trim() {
-                            on_save.call(current);
-                        }
-                        editing.set(None);
-                    },
-                }
-            } else {
-                div { class: "ebook-title-cell", "{display_value}" }
-                if let Some(err) = error.as_ref() {
-                    div { class: "error", "⚠ {err}" }
-                }
-            }
-        }
-    }
-}
-
-/// Inline-editable Authors cell. Unlike [`EditableCell`] (which hosts
-/// a single-line text input), the Authors cell renders the full
-/// [`ChipEditor`] *inside* the `<td>` when the row's editing signal
-/// matches `EditField::Authors`. The cell grows vertically to fit the
-/// chips + input + dropdown, matching the F5.9-lite design comp.
-///
-/// Display mode: comma-joined names. Hover: dashed amber outline.
-/// Click: swaps to chip editor (auto-focused) with the library-wide
-/// author pool surfaced in the dropdown. Escape exits edit mode.
-#[component]
-fn AuthorsCell(
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    authors_draft: Signal<Vec<String>>,
-    authors_text: String,
-    suggestions: ReadSignal<Vec<SuggestionItem>>,
-    on_change: EventHandler<Vec<String>>,
-) -> Element {
-    let mut editing = editing;
-    let is_editing = editing() == Some(EditField::Authors);
-    let active_class = if is_editing {
-        " ebook-cell-editing"
-    } else {
-        ""
-    };
-    let admin_class = if is_admin { " ebook-cell-editable" } else { "" };
-    let combined_class = format!("ebook-col-author{admin_class}{active_class}");
-
-    rsx! {
-        td {
-            class: "{combined_class}",
-            "data-testid": "ebook-cell-author",
-            onclick: move |e| {
-                if !is_admin {
-                    return;
-                }
-                e.stop_propagation();
-                if !is_editing {
-                    editing.set(Some(EditField::Authors));
-                }
-            },
-            if is_editing {
-                div { class: "ebook-cell-chip-host",
-                    ChipEditor {
-                        values: authors_draft,
-                        placeholder: "+ add author\u{2026}".to_string(),
-                        on_change,
-                        suggestions,
-                        show_avatar: false,
-                        aria_remove_prefix: "Remove".to_string(),
-                        testid_prefix: "ebook-cell-author".to_string(),
-                        autofocus: true,
-                        dropdown_header: "ADD AUTHOR".to_string(),
-                        on_close: move |_| {
-                            editing.set(None);
-                        },
-                    }
-                }
-            } else {
-                div { class: "ebook-title-cell", "{authors_text}" }
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Grid view
-// ---------------------------------------------------------------------------
-
-#[component]
-fn BookGrid(books: Vec<EbookMetadata>, server_url: String) -> Element {
-    rsx! {
-        div { class: "lib-grid", "data-testid": "lib-grid", role: "list",
-            for book in books.into_iter() {
-                GridTile {
-                    key: "{book.filename}",
-                    book: book,
-                    server_url: server_url.clone(),
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn GridTile(book: EbookMetadata, server_url: String) -> Element {
-    // Stable per-book uuid drives both detail-route URL and thumb URL
-    // (see `Route::BookDetail`).
-    let uuid = book.unique_identifier.clone().unwrap_or_default();
-    let display_title = book.title.as_deref().unwrap_or(&book.filename).to_string();
-    let tile_testid = format!("ebook-tile-{}", row_slug(&book.filename));
-    let authors = contributor_names(&book.creators);
-    let book_for_cover = book.clone();
-    let nav = use_navigator();
-
-    // Prefer the responsive `/api/thumbs/:uuid/{sm,md,lg}` endpoint over
-    // the raw `/api/covers/:uuid`: smaller payload (WebP, resized per
-    // slot) and the URL is server-prefixed so mobile picks up the right
-    // origin. Books with no cover fall back to the Atrium plate
-    // template.
-    let (thumb_src, thumb_srcset) = if book.cover_url.is_some() {
-        let base = format!("{server_url}/api/thumbs/{uuid}");
-        (
-            Some(format!("{base}/md")),
-            Some(format!("{base}/sm 160w, {base}/md 320w, {base}/lg 640w")),
-        )
-    } else {
-        (None, None)
-    };
-
-    let uuid_click = uuid.clone();
-    let uuid_key = uuid.clone();
-
-    rsx! {
-        a {
-            class: "cover-link lib-tile",
-            "data-testid": "{tile_testid}",
-            role: "listitem",
-            tabindex: "0",
-            aria_label: "Open details for {display_title}",
-            onclick: move |_| { nav.push(Route::BookDetail { uuid: uuid_click.clone() }); },
-            onkeydown: move |evt: Event<KeyboardData>| {
-                let key = evt.key();
-                if key == Key::Enter || key == Key::Character(" ".to_string()) {
-                    evt.prevent_default();
-                    nav.push(Route::BookDetail { uuid: uuid_key.clone() });
-                }
-            },
-            crate::components::atrium::Cover {
-                book: book_for_cover,
-                src_override: thumb_src,
-                srcset: thumb_srcset,
-                sizes: Some(
-                    "(max-width: 640px) 160px, (max-width: 1280px) 200px, 240px"
-                        .to_string(),
-                ),
-            }
-            div { class: "lib-tile-title", "{display_title}" }
-            if !authors.is_empty() {
-                div { class: "lib-tile-author", "{authors}" }
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Pure helpers — exercised by tests below.
-// ---------------------------------------------------------------------------
-
-fn contributor_names(list: &[Contributor]) -> String {
-    let mut out = String::new();
-    for (i, c) in list.iter().enumerate() {
-        if i > 0 {
-            out.push_str(", ");
-        }
-        out.push_str(&c.name);
-    }
-    out
-}
-
-fn primary_author_key(book: &EbookMetadata) -> String {
-    let c = book.creators.first();
-    let name = c
-        .map(|c| c.file_as.as_deref().unwrap_or(&c.name).to_string())
-        .unwrap_or_default();
-    name.to_ascii_lowercase()
-}
-
-fn title_key(book: &EbookMetadata) -> String {
-    let t = book.title.as_deref().unwrap_or(&book.filename);
-    t.to_ascii_lowercase()
-}
-
-/// Cached per-row sort key. We compute exactly one of these (matching the
-/// active [`SortKey`]) per book before sorting, then `sort_by` only borrows
-/// pre-built strings — no per-comparison allocation, no re-parsing of
-/// `series_index`. `series_index` is normalized to milli-units of an i64 so
-/// the whole struct is `Ord`-derivable (no f64 NaN issues).
-#[derive(PartialEq, Eq, PartialOrd, Ord)]
-struct RowKey {
-    /// Plain string axes (Title / Author / LastUpdated / NewestAdded).
-    /// `None` only for genuinely missing values; see [`cmp_with_missing_last`].
-    plain: Option<String>,
-    /// Series tuple: lowercased name + `series_index * 1000` rounded to i64.
-    series: Option<(String, i64)>,
-}
-
-fn row_key(book: &EbookMetadata, key: SortKey) -> RowKey {
-    match key {
-        SortKey::Title => RowKey {
-            plain: Some(title_key(book)),
-            series: None,
-        },
-        SortKey::Author => RowKey {
-            plain: Some(primary_author_key(book)),
-            series: None,
-        },
-        SortKey::Series => RowKey {
-            plain: None,
-            series: book.series.as_deref().filter(|s| !s.is_empty()).map(|s| {
-                let idx = book
-                    .series_index
-                    .as_deref()
-                    .and_then(|raw| raw.parse::<f64>().ok())
-                    .map(|f| (f * 1000.0).round() as i64)
-                    .unwrap_or(0);
-                (s.to_ascii_lowercase(), idx)
-            }),
-        },
-        SortKey::LastUpdated => RowKey {
-            plain: book.modified.clone(),
-            series: None,
-        },
-        SortKey::NewestAdded => RowKey {
-            plain: book.added_at.clone(),
-            series: None,
-        },
-    }
-}
-
-/// Compare two `Option<K>` values where missing always sorts last regardless
-/// of direction. Direction only flips ordering between two present values;
-/// `None` keeps a stable "last" position so reversing a desc sort doesn't
-/// shove un-timestamped or seriesless books to the top.
-fn cmp_with_missing_last<K: Ord>(a: Option<&K>, b: Option<&K>, dir: SortDir) -> Ordering {
-    match (a, b) {
-        (Some(x), Some(y)) => {
-            let ord = x.cmp(y);
-            if dir == SortDir::Desc {
-                ord.reverse()
-            } else {
-                ord
-            }
-        }
-        (Some(_), None) => Ordering::Less,
-        (None, Some(_)) => Ordering::Greater,
-        (None, None) => Ordering::Equal,
-    }
-}
-
-fn sort_books(books: Vec<EbookMetadata>, key: SortKey, dir: SortDir) -> Vec<EbookMetadata> {
-    let mut keyed: Vec<(RowKey, EbookMetadata)> =
-        books.into_iter().map(|b| (row_key(&b, key), b)).collect();
-    keyed.sort_by(|(ka, ba), (kb, bb)| {
-        let primary = match key {
-            SortKey::Series => cmp_with_missing_last(ka.series.as_ref(), kb.series.as_ref(), dir),
-            _ => cmp_with_missing_last(ka.plain.as_ref(), kb.plain.as_ref(), dir),
-        };
-        // Stable tiebreak on id, never reversed — keeps run-to-run order
-        // deterministic when the primary key matches.
-        primary.then(ba.id.cmp(&bb.id))
-    });
-    keyed.into_iter().map(|(_, b)| b).collect()
-}
-
-fn matches_filters(book: &EbookMetadata, filters: &ViewFilters) -> bool {
-    // Allocation-free membership checks: filter buckets are typically tiny
-    // (a handful of selected chips), so a nested `any().any()` is faster
-    // than building a fresh HashSet per book on every filter pass.
-    if !filters.authors.is_empty()
-        && !filters
-            .authors
-            .iter()
-            .any(|a| book.creators.iter().any(|c| &c.name == a))
-    {
-        return false;
-    }
-    if !filters.series.is_empty() {
-        let series = book.series.as_deref().unwrap_or("");
-        if !filters.series.iter().any(|s| s == series) {
-            return false;
-        }
-    }
-    if !filters.formats.is_empty()
-        && !filters
-            .formats
-            .iter()
-            .any(|f| book.formats.iter().any(|bf| bf.eq_ignore_ascii_case(f)))
-    {
-        return false;
-    }
-    if !filters.tags.is_empty()
-        && !filters
-            .tags
-            .iter()
-            .any(|t| book.subjects.iter().any(|s| s == t))
-    {
-        return false;
-    }
-    true
-}
-
-fn apply_filters(books: &[EbookMetadata], filters: &ViewFilters) -> Vec<EbookMetadata> {
-    if filters.is_empty() {
-        return books.to_vec();
-    }
-    books
-        .iter()
-        .filter(|b| matches_filters(b, filters))
-        .cloned()
-        .collect()
-}
-
-fn facet_counts(books: &[EbookMetadata]) -> FacetCounts {
-    let mut authors: BTreeMap<String, usize> = BTreeMap::new();
-    let mut series: BTreeMap<String, usize> = BTreeMap::new();
-    let mut formats: BTreeMap<String, usize> = BTreeMap::new();
-    let mut tags: BTreeMap<String, usize> = BTreeMap::new();
-    for book in books {
-        for c in &book.creators {
-            *authors.entry(c.name.clone()).or_default() += 1;
-        }
-        if let Some(s) = book.series.as_deref() {
-            if !s.is_empty() {
-                *series.entry(s.to_string()).or_default() += 1;
-            }
-        }
-        for fmt in &book.formats {
-            let key = fmt.trim().to_ascii_lowercase();
-            if !key.is_empty() {
-                *formats.entry(key).or_default() += 1;
-            }
-        }
-        for tag in &book.subjects {
-            if !tag.is_empty() {
-                *tags.entry(tag.clone()).or_default() += 1;
-            }
-        }
-    }
-    FacetCounts {
-        authors: sorted_facet(authors),
-        series: sorted_facet(series),
-        formats: sorted_facet(formats),
-        tags: sorted_facet(tags),
-    }
-}
-
-/// User-facing label for a normalized format key. Recognized formats get a
-/// friendly name (`"epub"` → `"ePub"`, `"m4b"` → `"Audiobook"`); anything
-/// else passes through upper-cased.
-fn format_display_label(key: &str) -> String {
-    match key {
-        "epub" => "ePub".to_string(),
-        "m4b" => "Audiobook".to_string(),
-        "pdf" => "PDF".to_string(),
-        "mp3" => "Audiobook (MP3)".to_string(),
-        other => other.to_ascii_uppercase(),
-    }
-}
-
-/// Short badge text for the table's Formats column. Stays compact so a row
-/// with two formats doesn't overflow the cell.
-fn format_badge_label(raw: &str) -> String {
-    raw.trim().to_ascii_uppercase()
-}
-
-fn sorted_facet(map: BTreeMap<String, usize>) -> Vec<(String, usize)> {
-    let mut v: Vec<(String, usize)> = map.into_iter().collect();
-    v.sort_by(|a, b| b.1.cmp(&a.1).then_with(|| a.0.cmp(&b.0)));
-    v
-}
-
-fn toggle_dir(d: SortDir) -> SortDir {
-    match d {
-        SortDir::Asc => SortDir::Desc,
-        SortDir::Desc => SortDir::Asc,
-    }
-}
-
-fn default_dir_for(key: SortKey) -> SortDir {
-    // "Newest Added" / "Last Updated" feel natural with newest first.
-    match key {
-        SortKey::NewestAdded | SortKey::LastUpdated => SortDir::Desc,
-        _ => SortDir::Asc,
-    }
-}
-
-const SORT_KEYS: [SortKey; 5] = [
-    SortKey::Title,
-    SortKey::Author,
-    SortKey::Series,
-    SortKey::LastUpdated,
-    SortKey::NewestAdded,
-];
-
-fn sort_key_value(key: SortKey) -> &'static str {
-    match key {
-        SortKey::Title => "title",
-        SortKey::Author => "author",
-        SortKey::Series => "series",
-        SortKey::LastUpdated => "last_updated",
-        SortKey::NewestAdded => "newest_added",
-    }
-}
-
-fn sort_key_label(key: SortKey) -> &'static str {
-    match key {
-        SortKey::Title => "Title",
-        SortKey::Author => "Author",
-        SortKey::Series => "Series",
-        SortKey::LastUpdated => "Last Updated",
-        SortKey::NewestAdded => "Newest Added",
-    }
-}
-
-fn sort_key_from_value(value: &str) -> Option<SortKey> {
-    match value {
-        "title" => Some(SortKey::Title),
-        "author" => Some(SortKey::Author),
-        "series" => Some(SortKey::Series),
-        "last_updated" => Some(SortKey::LastUpdated),
-        "newest_added" => Some(SortKey::NewestAdded),
-        _ => None,
-    }
-}
-
-/// Stable Playwright row id derived from the ebook's on-disk filename:
-/// strip directories and extension, lowercase, then collapse runs of
-/// non-alphanumeric ASCII characters into a single `-` (with leading and
-/// trailing dashes trimmed). The Playwright fixture table mirrors this
-/// derivation so each `FIXTURE_BOOKS[i].slug` matches the row's testid.
-fn row_slug(filename: &str) -> String {
-    let basename = filename.rsplit('/').next().unwrap_or(filename);
-    let stem = basename
-        .rsplit_once('.')
-        .map(|(s, _)| s)
-        .unwrap_or(basename);
-    let lower = stem.to_ascii_lowercase();
-    let mut out = String::with_capacity(lower.len());
-    let mut last_was_dash = true;
-    for ch in lower.chars() {
-        if ch.is_ascii_alphanumeric() {
-            out.push(ch);
-            last_was_dash = false;
-        } else if !last_was_dash {
-            out.push('-');
-            last_was_dash = true;
-        }
-    }
-    while out.ends_with('-') {
-        out.pop();
-    }
-    out
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use omnibus_shared::Contributor;
-
-    #[allow(clippy::too_many_arguments)]
-    fn book(
-        id: i64,
-        filename: &str,
-        title: Option<&str>,
-        authors: &[(&str, Option<&str>)],
-        series: Option<(&str, &str)>,
-        modified: Option<&str>,
-        added_at: Option<&str>,
-        subjects: &[&str],
-    ) -> EbookMetadata {
-        EbookMetadata {
-            id,
-            filename: filename.into(),
-            title: title.map(Into::into),
-            creators: authors
-                .iter()
-                .map(|(name, file_as)| Contributor {
-                    name: (*name).into(),
-                    role: None,
-                    file_as: file_as.map(Into::into),
-                    id: None,
-                })
-                .collect(),
-            series: series.map(|(s, _)| s.into()),
-            series_index: series.map(|(_, i)| i.into()),
-            modified: modified.map(Into::into),
-            added_at: added_at.map(Into::into),
-            subjects: subjects.iter().map(|s| (*s).to_string()).collect(),
-            ..Default::default()
-        }
-    }
-
-    fn ids(books: &[EbookMetadata]) -> Vec<i64> {
-        books.iter().map(|b| b.id).collect()
-    }
-
-    // --- row_slug ---
-
-    #[test]
-    fn row_slug_lowercases_and_strips_extension() {
-        assert_eq!(row_slug("Alpha.epub"), "alpha");
-    }
-    #[test]
-    fn row_slug_collapses_runs_of_non_alphanumerics() {
-        assert_eq!(row_slug("Beta in the Series.epub"), "beta-in-the-series");
-    }
-    #[test]
-    fn row_slug_uses_basename_for_nested_paths() {
-        assert_eq!(row_slug("series/vol1/Deep Book.epub"), "deep-book");
-    }
-    #[test]
-    fn row_slug_trims_trailing_dashes() {
-        assert_eq!(row_slug("weird---name!!!.epub"), "weird-name");
-    }
-    #[test]
-    fn row_slug_handles_filename_without_extension() {
-        assert_eq!(row_slug("plain"), "plain");
-    }
-
-    // --- sort_books ---
-
-    fn sample() -> Vec<EbookMetadata> {
-        vec![
-            book(
-                1,
-                "alpha.epub",
-                Some("Alpha"),
-                &[("Tolkien, J.R.R.", Some("Tolkien, J.R.R."))],
-                Some(("Foundation", "1")),
-                Some("2024-01-01T00:00:00"),
-                Some("2025-03-10T00:00:00"),
-                &["Fantasy"],
-            ),
-            book(
-                2,
-                "beta.epub",
-                Some("Beta"),
-                &[("Asimov, Isaac", Some("Asimov, Isaac"))],
-                Some(("Foundation", "2")),
-                Some("2024-06-01T00:00:00"),
-                Some("2025-01-05T00:00:00"),
-                &["Sci-Fi"],
-            ),
-            book(
-                3,
-                "gamma.epub",
-                Some("Gamma"),
-                &[("Le Guin, Ursula", Some("Le Guin, Ursula"))],
-                None,
-                Some("2023-01-01T00:00:00"),
-                Some("2025-02-20T00:00:00"),
-                &["Fantasy", "Sci-Fi"],
-            ),
-        ]
-    }
-
-    #[test]
-    fn sorts_by_title_asc_and_desc() {
-        let s = sample();
-        let asc = sort_books(s.clone(), SortKey::Title, SortDir::Asc);
-        assert_eq!(ids(&asc), vec![1, 2, 3]);
-        let desc = sort_books(s, SortKey::Title, SortDir::Desc);
-        assert_eq!(ids(&desc), vec![3, 2, 1]);
-    }
-
-    #[test]
-    fn sorts_by_author_asc() {
-        let s = sample();
-        let asc = sort_books(s, SortKey::Author, SortDir::Asc);
-        // Asimov < Le Guin < Tolkien
-        assert_eq!(ids(&asc), vec![2, 3, 1]);
-    }
-
-    #[test]
-    fn sorts_by_series_grouping_with_index_then_pushes_seriesless_last() {
-        let s = sample();
-        let asc = sort_books(s, SortKey::Series, SortDir::Asc);
-        // Foundation #1 (id 1), Foundation #2 (id 2), then no-series (id 3).
-        assert_eq!(ids(&asc), vec![1, 2, 3]);
-    }
-
-    #[test]
-    fn sorts_by_last_updated_desc_picks_most_recent_first() {
-        let s = sample();
-        let desc = sort_books(s, SortKey::LastUpdated, SortDir::Desc);
-        // beta 2024-06 > alpha 2024-01 > gamma 2023
-        assert_eq!(ids(&desc), vec![2, 1, 3]);
-    }
-
-    #[test]
-    fn sorts_by_newest_added_desc() {
-        let s = sample();
-        let desc = sort_books(s, SortKey::NewestAdded, SortDir::Desc);
-        // alpha 2025-03 > gamma 2025-02 > beta 2025-01
-        assert_eq!(ids(&desc), vec![1, 3, 2]);
-    }
-
-    #[test]
-    fn missing_timestamps_always_sort_last_even_on_desc() {
-        // Two timestamped books + one with no `modified` value. In descending
-        // order the most-recent timestamp comes first, but the missing-value
-        // book stays at the end (it doesn't get flipped to the top by the
-        // direction reversal).
-        let books = vec![
-            book(
-                1,
-                "old.epub",
-                Some("Old"),
-                &[],
-                None,
-                Some("2024-01-01T00:00:00"),
-                None,
-                &[],
-            ),
-            book(
-                2,
-                "new.epub",
-                Some("New"),
-                &[],
-                None,
-                Some("2025-01-01T00:00:00"),
-                None,
-                &[],
-            ),
-            book(
-                3,
-                "missing.epub",
-                Some("Missing"),
-                &[],
-                None,
-                None,
-                None,
-                &[],
-            ),
-        ];
-        let desc = sort_books(books.clone(), SortKey::LastUpdated, SortDir::Desc);
-        assert_eq!(ids(&desc), vec![2, 1, 3]);
-        let asc = sort_books(books, SortKey::LastUpdated, SortDir::Asc);
-        assert_eq!(ids(&asc), vec![1, 2, 3]);
-    }
-
-    #[test]
-    fn series_sort_keeps_seriesless_last_in_desc_too() {
-        let s = sample();
-        let desc = sort_books(s, SortKey::Series, SortDir::Desc);
-        // Foundation #2 (id 2) → Foundation #1 (id 1) → seriesless gamma (id 3)
-        // last regardless of direction.
-        assert_eq!(ids(&desc), vec![2, 1, 3]);
-    }
-
-    // --- apply_filters ---
-
-    #[test]
-    fn empty_filters_returns_all_books() {
-        let s = sample();
-        let out = apply_filters(&s, &ViewFilters::default());
-        assert_eq!(ids(&out), vec![1, 2, 3]);
-    }
-
-    #[test]
-    fn single_facet_or_within_group() {
-        let s = sample();
-        let f = ViewFilters {
-            authors: vec!["Tolkien, J.R.R.".into(), "Asimov, Isaac".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&s, &f);
-        assert_eq!(ids(&out), vec![1, 2]);
-    }
-
-    #[test]
-    fn multi_facet_and_across_groups() {
-        let s = sample();
-        let f = ViewFilters {
-            authors: vec!["Tolkien, J.R.R.".into(), "Asimov, Isaac".into()],
-            series: vec!["Foundation".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&s, &f);
-        // alpha is Tolkien + Foundation, beta is Asimov + Foundation, gamma
-        // is Le Guin with no series — so only alpha + beta survive the AND
-        // across (authors, series).
-        assert_eq!(ids(&out), vec![1, 2]);
-    }
-
-    #[test]
-    fn series_filter_excludes_books_with_no_series() {
-        let s = sample();
-        let f = ViewFilters {
-            series: vec!["Foundation".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&s, &f);
-        assert_eq!(ids(&out), vec![1, 2]);
-    }
-
-    // --- facet_counts ---
-
-    #[test]
-    fn facet_counts_orders_by_count_desc_then_name() {
-        let s = sample();
-        let f = facet_counts(&s);
-        // Series: Foundation present once with count 2
-        assert_eq!(f.series, vec![("Foundation".into(), 2)]);
-        // Authors: each unique once
-        assert_eq!(f.authors.len(), 3);
-    }
-
-    #[test]
-    fn facet_counts_skips_empty_series_strings() {
-        let mut b = sample();
-        b[0].series = Some(String::new());
-        let f = facet_counts(&b);
-        assert!(f.series.iter().all(|(s, _)| !s.is_empty()));
-    }
-
-    // --- format filter ---
-
-    fn with_formats(mut b: EbookMetadata, formats: &[&str]) -> EbookMetadata {
-        b.formats = formats.iter().map(|s| (*s).to_string()).collect();
-        b
-    }
-
-    #[test]
-    fn format_counts_normalize_case_insensitively() {
-        let books = vec![
-            with_formats(sample()[0].clone(), &["EPUB"]),
-            with_formats(sample()[1].clone(), &["epub", "m4b"]),
-            with_formats(sample()[2].clone(), &["M4B"]),
-        ];
-        let f = facet_counts(&books);
-        let formats: std::collections::HashMap<_, _> = f.formats.into_iter().collect();
-        assert_eq!(formats.get("epub").copied(), Some(2));
-        assert_eq!(formats.get("m4b").copied(), Some(2));
-    }
-
-    #[test]
-    fn empty_formats_filter_keeps_all_books() {
-        let books = vec![
-            with_formats(sample()[0].clone(), &["EPUB"]),
-            with_formats(sample()[1].clone(), &["m4b"]),
-        ];
-        let out = apply_filters(&books, &ViewFilters::default());
-        assert_eq!(ids(&out), vec![1, 2]);
-    }
-
-    #[test]
-    fn format_filter_or_within_bucket() {
-        let books = vec![
-            with_formats(sample()[0].clone(), &["epub"]),
-            with_formats(sample()[1].clone(), &["m4b"]),
-            with_formats(sample()[2].clone(), &["pdf"]),
-        ];
-        let f = ViewFilters {
-            formats: vec!["epub".into(), "m4b".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&books, &f);
-        assert_eq!(ids(&out), vec![1, 2]);
-    }
-
-    #[test]
-    fn format_filter_matches_case_insensitively() {
-        // A book whose persisted format string is upper-case "EPUB" should
-        // still match a filter chip whose normalized key is "epub".
-        let books = vec![with_formats(sample()[0].clone(), &["EPUB"])];
-        let f = ViewFilters {
-            formats: vec!["epub".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&books, &f);
-        assert_eq!(ids(&out), vec![1]);
-    }
-
-    #[test]
-    fn format_filter_intersects_with_other_facets() {
-        // Tolkien wrote `alpha` (Fantasy + Foundation) and only that book is
-        // EPUB — a Tolkien + EPUB filter should leave just alpha.
-        let books = vec![
-            with_formats(sample()[0].clone(), &["epub"]),
-            with_formats(sample()[1].clone(), &["m4b"]),
-        ];
-        let f = ViewFilters {
-            authors: vec!["Tolkien, J.R.R.".into()],
-            formats: vec!["epub".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&books, &f);
-        assert_eq!(ids(&out), vec![1]);
-    }
-
-    #[test]
-    fn format_display_label_friendly_names() {
-        assert_eq!(format_display_label("epub"), "ePub");
-        assert_eq!(format_display_label("m4b"), "Audiobook");
-        assert_eq!(format_display_label("pdf"), "PDF");
-        // Unknown formats fall through upper-cased.
-        assert_eq!(format_display_label("azw3"), "AZW3");
-    }
 
     #[test]
     fn short_path_returns_last_segment() {
         assert_eq!(short_path("/Users/ek/books"), "books");
         assert_eq!(short_path("/Users/ek/books/"), "books");
         assert_eq!(short_path("relative"), "relative");
-    }
-
-    // --- additional filter / facet / sort edge cases (#215) ---
-
-    #[test]
-    fn tag_filter_is_or_within_bucket_not_and() {
-        // tags is OR within the bucket: selecting two subjects must keep books
-        // carrying *either* one, never require *both* (AND would wrongly drop
-        // the single-tag books). alpha=Fantasy, beta=Sci-Fi, gamma=both.
-        let s = sample();
-        let both = ViewFilters {
-            tags: vec!["Fantasy".into(), "Sci-Fi".into()],
-            ..Default::default()
-        };
-        // OR keeps all three; AND would have returned only gamma (id 3).
-        assert_eq!(ids(&apply_filters(&s, &both)), vec![1, 2, 3]);
-
-        // A single selected tag still matches every book carrying it.
-        let one = ViewFilters {
-            tags: vec!["Fantasy".into()],
-            ..Default::default()
-        };
-        assert_eq!(ids(&apply_filters(&s, &one)), vec![1, 3]);
-    }
-
-    #[test]
-    fn multi_format_filter_is_or_within_bucket_not_and() {
-        // Guards against the bug called out in #215: selecting two formats must
-        // include books matching *either* (OR within the formats bucket), never
-        // require a book to carry *both* (AND would wrongly exclude single-format
-        // books). Each book here has exactly one format.
-        let books = vec![
-            with_formats(sample()[0].clone(), &["epub"]),
-            with_formats(sample()[1].clone(), &["m4b"]),
-            with_formats(sample()[2].clone(), &["pdf"]),
-        ];
-        let f = ViewFilters {
-            formats: vec!["epub".into(), "m4b".into()],
-            ..Default::default()
-        };
-        let out = apply_filters(&books, &f);
-        // OR semantics keep both single-format epub and m4b books; AND would
-        // have dropped both and returned an empty list.
-        assert_eq!(ids(&out), vec![1, 2]);
-    }
-
-    #[test]
-    fn format_facet_counts_tally_across_full_unfiltered_list() {
-        // facet_counts must accumulate over every book it is given (the full,
-        // unfiltered list) — duplicates across books increment the same key.
-        let books = vec![
-            with_formats(sample()[0].clone(), &["epub"]),
-            with_formats(sample()[1].clone(), &["epub", "m4b"]),
-            with_formats(sample()[2].clone(), &["m4b"]),
-        ];
-        let f = facet_counts(&books);
-        let formats: std::collections::HashMap<_, _> = f.formats.into_iter().collect();
-        assert_eq!(formats.get("epub").copied(), Some(2));
-        assert_eq!(formats.get("m4b").copied(), Some(2));
-        assert_eq!(formats.len(), 2);
-    }
-
-    #[test]
-    fn sort_by_title_is_stable_on_equal_keys_via_id_tiebreak() {
-        // Two books share the same title; the id tiebreak keeps a deterministic
-        // order regardless of input order or sort direction.
-        let mut a = book(5, "a.epub", Some("Same"), &[], None, None, None, &[]);
-        let mut b = book(2, "b.epub", Some("Same"), &[], None, None, None, &[]);
-        a.title = Some("Same".into());
-        b.title = Some("Same".into());
-        let asc = sort_books(vec![a.clone(), b.clone()], SortKey::Title, SortDir::Asc);
-        assert_eq!(ids(&asc), vec![2, 5]);
-        // Reversed direction does not reverse the tiebreak: ids stay ascending.
-        let desc = sort_books(vec![a, b], SortKey::Title, SortDir::Desc);
-        assert_eq!(ids(&desc), vec![2, 5]);
-    }
-
-    #[test]
-    fn empty_filters_returns_full_list_unchanged_in_original_order() {
-        // Acceptance (4): an empty ViewFilters yields the input list verbatim.
-        let s = sample();
-        let out = apply_filters(&s, &ViewFilters::default());
-        assert_eq!(ids(&out), ids(&s));
     }
 }
